Move pool data hook and calculation to package
