+++ v3/lib/usePoolData/usePoolData.tsx
@@ -1,12 +1,12 @@
 import { useQuery } from '@tanstack/react-query';
-import { calculateMarketPnl } from '../utils/calculations';
-import { formatGraphBigDecimal, getSubgraphUrl } from '../utils/subgraph';
+import { calculateMarketPnl } from '@snx-v3/calculations';
+import { getSubgraphUrl } from '@snx-v3/Constants';
 import { z } from 'zod';
 import { wei } from '@synthetixio/wei';
 import { useNetwork } from '@snx-v3/useBlockchain';
 
-const GraphBigIntSchema = z.string().transform((src) => formatGraphBigDecimal(src));
-const GraphBigDecimalSchema = z.string().transform((src) => formatGraphBigDecimal(src));
+const GraphBigIntSchema = z.string().transform((src) => wei(src));
+const GraphBigDecimalSchema = z.string().transform((src) => wei(src));
 
 export const MarketSnapshotByWeekSchema = z
   .object({
@@ -111,7 +111,7 @@ const getPoolData = async (chainName: string, id: string) => {
   return PoolDataResultSchema.parse(json);
 };
 
-export const useGetPoolData = (poolId?: string) => {
+export const usePoolData = (poolId?: string) => {
   const network = useNetwork();
 
   return useQuery({
